Superclasse GameCreation
Realizzata ereditarietà tra classi GameCreation (superclasse) e CreateSingleGameController e CreateTournamentController (sottoclassi che estendono GameCreation)
