+++ GiocoSPACCA/Informazioni_Partite/PLAYERS_REGISTER.csv
@@ -3,8 +3,8 @@ zano,6
 cris,7
 francesca,38
 franceschinapotentissima,3
-fula,21
+fula,25
 bolognino,51
 tommtomm,16
 enzo,3
-agnes,50
+agnes,52
